pyproject.toml
[tool.poetry]
name = "tg_autoreply_bot"
version = "0.1.0"
description = "Auto reply on the messages if user is not online."
authors = ["Organ13at0r <Organ13at0r_official@proton.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Telethon = "^1.25.4"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
